+++ src/main/kotlin/io/unthrottled/doki/themes/ThemeManager.kt
@@ -1,16 +1,16 @@
 package io.unthrottled.doki.themes
 
 import com.intellij.openapi.Disposable
-import com.intellij.openapi.components.ServiceManager
+import com.intellij.openapi.application.ApplicationManager
 import java.util.Optional
 import javax.swing.UIManager
 
 interface ThemeManager : Disposable {
   companion object {
-    const val DEFAULT_THEME_NAME = "Re:Zero: Rem"
+    const val DEFAULT_THEME_NAME = "Slime: Rimiru Tempest"
 
     val instance: ThemeManager
-      get() = ServiceManager.getService(ThemeManager::class.java)
+      get() = ApplicationManager.getApplication().getService(ThemeManager::class.java)
   }
 
   val isCurrentThemeDoki: Boolean